Moved MultiKeyJSON and Base64Encode into separate files
Candidates for moving into Bricks
